+++ files/private-chef-cookbooks/private-chef/recipes/oc_bifrost.rb
@@ -23,8 +23,6 @@ directory "/opt/opscode/sv/opscode-authz" do
 end
 # END AUTHZ CLEANUP
 
-owner = node['private_chef']['user']['username']
-group = owner
 
 oc_bifrost_dir = node['private_chef']['oc_bifrost']['dir']
 oc_bifrost_bin_dir = File.join(oc_bifrost_dir, "bin")
@@ -39,8 +37,9 @@ oc_bifrost_sasl_log_dir = File.join(oc_bifrost_log_dir, "sasl")
   oc_bifrost_sasl_log_dir
 ].each do |dir_name|
   directory dir_name do
-    owner owner
-    mode "0700"
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end
@@ -62,8 +61,8 @@ oc_bifrost_config = File.join(oc_bifrost_etc_dir, "sys.config")
 
 template oc_bifrost_config do
   source "oc_bifrost.config.erb"
-  owner owner
-  group group
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "644"
   variables(node['private_chef']['oc_bifrost'].to_hash)
   notifies :restart, 'runit_service[oc_bifrost]' unless backend_secondary?