+++ files/private-chef-cookbooks/private-chef/recipes/opscode-account.rb
@@ -17,8 +17,9 @@ private_chef_account_log_dir = node['private_chef']['opscode-account']['log_dire
   private_chef_account_log_dir
 ].each do |dir_name|
   directory dir_name do
-    owner node['private_chef']['user']['username']
-    mode '0700'
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end