+++ files/private-chef-cookbooks/private-chef/recipes/opscode-certificate.rb
@@ -5,8 +5,6 @@
 # All Rights Reserved
 #
 
-owner = node['private_chef']['user']['username']
-group = owner
 
 opscode_certificate_dir = node['private_chef']['opscode-certificate']['dir']
 opscode_certificate_etc_dir = File.join(opscode_certificate_dir, "etc")
@@ -17,8 +15,9 @@ opscode_certificate_log_dir = node['private_chef']['opscode-certificate']['log_d
   opscode_certificate_log_dir,
 ].each do |dir_name|
   directory dir_name do
-    owner owner
-    mode "0700"
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end
@@ -31,8 +30,8 @@ certificate_config = File.join(opscode_certificate_etc_dir, "certificate.config"
 
 template certificate_config do
   source "certgen_web.config.erb"
-  owner owner
-  group group
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "644"
   variables(node['private_chef']['opscode-certificate'].to_hash)
   notifies :restart, 'runit_service[opscode-certificate]'