+++ files/private-chef-cookbooks/private-chef/recipes/opscode-org-creator.rb
@@ -4,8 +4,6 @@
 #
 # All Rights Reserved
 
-owner = node['private_chef']['user']['username']
-group = owner
 
 opscode_org_creator_dir = node['private_chef']['opscode-org-creator']['dir']
 opscode_org_creator_etc_dir = File.join(opscode_org_creator_dir, "etc")
@@ -18,8 +16,9 @@ opscode_org_creator_log_sasl_dir = File.join(node['private_chef']['opscode-org-c
   opscode_org_creator_log_sasl_dir
 ].each do |dir_name|
   directory dir_name do
-    owner owner
-    mode "0700"
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end
@@ -28,8 +27,8 @@ org_creator_config = File.join(opscode_org_creator_etc_dir, "app.config")
 
 template org_creator_config do
   source "opscode-org-creator.config.erb"
-  owner owner
-  group group
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "644"
   variables(node['private_chef']['opscode-org-creator'].to_hash)
   notifies :restart, 'runit_service[opscode-org-creator]' if is_data_master?
@@ -45,8 +44,8 @@ end
 
 template "/opt/opscode/embedded/service/opscode-org-creator/rel/org_app/bin/org_app" do
   source "org_app.erb"
-  owner owner
-  group group
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "0755"
   notifies :restart, 'runit_service[opscode-org-creator]' if is_data_master?
 end