+++ files/private-chef-cookbooks/private-chef/recipes/opscode-solr.rb
@@ -15,8 +15,9 @@ solr_log_dir = node['private_chef']['opscode-solr']['log_directory']
 
 [ solr_dir, solr_etc_dir, solr_data_dir, solr_home_dir, solr_log_dir ].each do |dir_name|
   directory dir_name do
-    owner node['private_chef']['user']['username']
-    mode '0700'
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end
@@ -48,7 +49,7 @@ execute "cp -R /opt/opscode/embedded/service/opscode-solr/jetty #{File.dirname(s
   notifies(:restart, "runit_service[opscode-solr]") if is_data_master?
 end
 
-execute "chown -R #{node['private_chef']['user']['username']} #{solr_dir}" do
+execute "chown -R #{OmnibusHelper.new(node).ownership['owner']} #{solr_dir}" do
   not_if { File.exists?(solr_installed_file) }
 end
 
@@ -61,7 +62,8 @@ file solr_installed_file do
 end
 
 template File.join(solr_jetty_dir, "etc", "jetty.xml") do
-  owner node['private_chef']['user']['username']
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "0644"
   source "jetty.xml.erb"
   variables(node['private_chef']['opscode-solr'].to_hash.merge(node['private_chef']['logs'].to_hash))
@@ -69,7 +71,8 @@ template File.join(solr_jetty_dir, "etc", "jetty.xml") do
 end
 
 template File.join(solr_home_dir, "conf", "solrconfig.xml") do
-  owner node['private_chef']['user']['username']
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode "0644"
   source "solrconfig.xml.erb"
   variables(node['private_chef']['opscode-solr'].to_hash)